feat: Add Prompt expansion script
This commit adds the new script "Prompt expansion" to the UI. The script
expands python list comprehensions in prompts by evaluating and
expanding expressions enclosed in brackets. It supports multiple
brackets in a single prompt as well as positive and negative prompts,
resulting in products of combinations. The script provides options for
returning multiline or prompt-based outputs.